fix: split monorepo nuxt test out of check_cli.rs; allow optional virtualTs in snapshot baseline check
